baseline/api_gateway/requirements.txt
fastapi
httpx
python-multipart
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
alembic
celery
PyJWT[crypto]
pypdf
python-docx
